Refactor: 그룹 모델뷰 생성
